src/bot/scripts/bases/requirements.txt
pandas>=1.5.0
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
openpyxl>=3.0.0
pathlib
datetime
